Unify cursor + savepoint suites as prolly_txn_stress_test
Fold cursor_merge_stress_test and savepoint_txn_matrix_test into a
single gating binary that covers TREE+MUT cursor merge and the nested
SAVEPOINT × mutmap × DDL × VC matrix. One CI loop job
(prolly-txn-stress) under DOLTLITE_PROLLY_CHECK=1.
