+++ test/run_c_tests.sh
@@ -15,8 +15,7 @@ GATING=(
   prolly_chunker_boundary_test
   scoped_refs_push_test
   concurrent_stress_test
-  cursor_merge_stress_test
-  savepoint_txn_matrix_test
+  prolly_txn_stress_test
   vc_concurrency_test
   vc_ref_mutation_stress_test
   multi_process_test
